Fixed problems with encryption when combined with compression or sparse files.  Unfortunately because of this bug all previous encrypted backups can't be restored.
Fixed problem in bfgets with Windows and Mac end of lines.

Fixed bug in Windows version of close_bpipe() that caused programs like bsmtp() to hang waiting for the pipe to be closed.

Fixed binary file type problems in open_bpipe().

Fixed a number of bugs in the Windows version of the mysql DB scripts.