On vacation I am having a little fun !!
Storage status, dir status and client status pages all now have separate timers 
which are visible and configurable from within the page.  The user can see the
countdown of the timer.

Add a tab widget to the client status for a little different look containing
the heading and running outputs.  Changing to a different tab auto refreshes
the widget that becomes in front like with the storage status.

In the client status page make the table right click to a refresh option.

In both client and storage pages, if the right click option for a status window
is chosen, a test is done to see if that staus already exists.  If so, then
bring that status window to the current window instead of creating a repeat 
instance of the class

In the storage status page, miscelaneous tab - add two more buttons to perform
commands for label and release identical to the right click options from the
storage page.

If a status window for a specific client or storage is open when bat is closed,
a settings is written and if the storage or client page is populated the next
time bat is opened, the status page for the storage or client will automatically
be created.

In the main menu, add an option to execute "reload" AND ALSO repopulate the
lists for an easy way to do both after modifying bacula-dir.conf

With each status page having it's own timer, the common timer from prefs is no
longer required

The joblist page now has a column for pool.  A dropdown for selecting based on
pool has also been added.

Restoretree: The bug that was issued for a directory not being restored has
been resolved.  I thought this would take a while.  Turns out a query just
needed one line removed.